Feat/#4: 회원가입 및 로그인 기능 추가

+++ src/main/java/inha/git/mapping/domain/id/FoundingRecommendId.java
@@ -18,7 +18,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class FoundingRecommendId implements Serializable {
 
-    private Long projectId;
-    private Long userId;
+    private Integer projectId;
+    private Integer userId;
 
 }
